fix: Change order of role to avoid restarting slurmd
cod_slurm templates the slurm config and freezes all overwrites to slurm.conf by
CMD. The desired config should be available by the start of compute node build
not after, so that they will picked up during the build.
This way slurmd restart is not needed on the compute node.