feat(account_app): separate account app to run on dedicated node
The motivation for this change is to run the account app as an independent OpenStack instance on its own dedicated node. Currently, the account app is combined with the Open OnDemand (OOD) node

- Copied user_reg_cloud role to new account_app role for isolation
- Retained essential tasks for account app deployment and setup
- Removed tasks and files tied to OpenOnDemand (OOD) and Apache as they
  do not run on the account app node
         user-reg_conf.j2, user-reg_conf_shib.j2 and user_auth_py.j2
- Updated hosts and added acctapp.yaml to define account app hosts inventory
- Added logging to account_app role tasks

The reason for copying rather than moving the role is that the existing roles ood_user_reg_cloud and ood_user_reg_ops are tightly tied to the OOD build and can be disabled and removed in a future sprint. Creating a new dedicated role for the account app isolates changes and simplifies incremental migration without disrupting the current OOD deployment.